Restructure of project organization 2
