+++ tuva_dqi/pages/page__report_card.py
@@ -6,14 +6,12 @@ import dash_bootstrap_components as dbc
 import pandas as pd
 import plotly.express as px
 from dash import Input, Output, callback, dcc, html
-
-# Import helper functions from analytics page
-from tuva_dqi.utils import (
-    create_chart,
+from tuva_dqi.pages.charts import create_chart
+from tuva_dqi.pages.db import get_db_connection
+from tuva_dqi.pages.services import (
     get_all_tests,
     get_available_charts,
     get_data_quality_grade,
-    get_db_connection,
     get_last_test_run_time,
     get_mart_test_summary,
     get_outstanding_errors,
